move usermodule in vhus
