+++ tatk/policy/mle/camrest/mle.py
@@ -46,7 +46,7 @@ class MLP(Policy):
         pass
     
     def load(self, filename):
-        policy_mdl = filename + '_mle.pol.mdl'
+        policy_mdl = os.path.join(os.path.dirname(os.path.abspath(__file__)), filename + '_mle.pol.mdl')
         if os.path.exists(policy_mdl):
             self.policy.load_state_dict(torch.load(policy_mdl))
             print('<<dialog policy>> loaded checkpoint from file: {}'.format(policy_mdl))
